Module Provider_test.Interface

module Directory_reader : sig ... end

A directory reader is an interface that is able to list all the entries contained in a directory in the file system.

module File_reader : sig ... end

An interface that is able to load the contents of a file from the file system.

module Float_printer : sig ... end

An interface that is able to print floats.

module Int_printer : sig ... end

An interface that is able to print ints.