Module type Provider_interface.S

type t

The implementation may be based on any type it wants, that's the whole point.

val readdir : t -> path:Base.string -> Base.string Base.list

Note that not all the interface exported by the module is required for the implementer, indeed only the minimal set of methods upon which we can build all the others. In this case, all that is required is the ability to list entries from a given directory.