Module Interface.File_reader

An interface that is able to load the contents of a file from the file system.

The structure of this file is very similar to the Directory_reader module, thus is not documented in details. Refer to Directory_reader for more.

type tag = [

| `File_reader

]

type 'a t = [> tag ] as 'a Provider.t

val load : _ t -> path:Base.string -> Base.string

module Provider_interface : sig ... end