Module Providers.Eio_reader

Eio_reader is a provider for the Interface.Directory_reader and Interface.File_reader interfaces, based on Eio.

It is a thin wrapper around Eio.Path, meant to demonstrate how to implement a provider for a given interface.

Unix_reader is another provider for the Interface.Directory_reader interface.

type t

Depending on the needs, the type t may be exposed or not. It can be anything required by the implementation.

val make : env:< fs : _ Eio.Path.t.. > -> [ Interface.Directory_reader.tag | Interface.File_reader.tag ] Provider.t

Somehow there must be a way exposed to construct a provider, decorated with the tags that the provider implements. Here, the construct [ `Directory_reader | `File_reader ] means that this provider implements both interfaces: Interface.Directory_reader and Interface.File_reader.

module Impl : sig ... end

In this specific example, we chose to expose the signature of the provider's implementation. This is not something that is usually required, since make already provides a way to build a provider. This is only done here for the sake of the tests, as we show different ways an interface can be built based on various parts of its implementation.