Module Providers.Num_printer

Num_printer is a provider for the Interface.Int_printer and Interface.Float_printer interfaces.

The structure of this file is very similar to the Eio_reader module, thus is not documented in details. Refer to Eio_reader for more.

type t = Base.unit

In this case we decided to expose the type t and interface, to demonstrate how to override a particular class of the provided interface. See test__override.ml.

val interface : (t, [ Interface.Int_printer.tag | Interface.Float_printer.tag ]) Provider.Interface.t

val make : t -> [ Interface.Int_printer.tag | Interface.Float_printer.tag ] Provider.t

If you simply wish to use this provider without overrides, use make ().